% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpparma_bothproducts}
\alias{rcpparma_bothproducts}
\title{Return both the inner and outer product}
\usage{
rcpparma_bothproducts(x)
}
\arguments{
\item{x}{column vector}
}
\value{
list with both the inner and outer product
}
\description{
Return a list with both the inner and outer product
}

