% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpparma_hello_world}
\alias{rcpparma_hello_world}
\title{Hello World}
\usage{
rcpparma_hello_world()
}
\value{
matrix
}
\description{
simple example of creating two matrices and returning the result of an
operation on them
}

