% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpparma_innerproduct}
\alias{rcpparma_innerproduct}
\title{Inner product}
\usage{
rcpparma_innerproduct(x)
}
\arguments{
\item{x}{column vector}
}
\value{
scalar with the inner product
}
\description{
another simple example: inner product of a vector,
returning a scalar
}

