% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpparma_outerproduct}
\alias{rcpparma_outerproduct}
\title{Outer product}
\usage{
rcpparma_outerproduct(x)
}
\arguments{
\item{x}{column vector}
}
\value{
matrix with the outer product
}
\description{
another simple example: outer product of a vector,
returning a matrix
}

